+++ server/src/Server/User.js
@@ -3,9 +3,7 @@
 const { GameObject } = require('ud-viz/src/Game/Game');
 const Game = require('ud-viz/src/Game/Game');
 const Constants = Game.Components.Constants;
-const Command = Game.Command;
 const WorldState = Game.WorldState;
-const WorldThread = require('./WorldThread');
 
 const UserModule = class User {
   constructor(uuid, socket, data, isGuest = false) {
@@ -62,43 +60,6 @@ const UserModule = class User {
     //assign
     this.thread = thread;
     this.lastState = null;
-    this.socket.removeAllListeners(Constants.WEBSOCKET.MSG_TYPES.COMMANDS);
-    this.socket.removeAllListeners(
-      Constants.WEBSOCKET.MSG_TYPES.ADD_GAMEOBJECT
-    );
-
-    if (thread) {
-      //cmds are now sent to the new thread
-      const _this = this;
-      this.socket.on(
-        Constants.WEBSOCKET.MSG_TYPES.COMMANDS,
-        function (cmdsJSON) {
-          const commands = [];
-
-          //parse
-          cmdsJSON.forEach(function (cmdJSON) {
-            const command = new Command(cmdJSON);
-
-            if (command.getUserID() == _this.getUUID()) {
-              //security so another client cant control another avatar
-              commands.push(command);
-            }
-          });
-
-          _this.thread.post(WorldThread.MSG_TYPES.COMMANDS, commands);
-        }
-      );
-
-      //add go
-      this.socket.on(
-        Constants.WEBSOCKET.MSG_TYPES.ADD_GAMEOBJECT,
-        function (goJSON) {
-          _this.thread.post(WorldThread.MSG_TYPES.ADD_GAMEOBJECT, {
-            gameObject: goJSON,
-          });
-        }
-      );
-    }
   }
 
   getUUID() {
